long

+++ example.py
@@ -8,7 +8,7 @@ model = Kosmos()
 batch_size = 4
 dummy_images = torch.randn(batch_size, 3, 224, 224)  # Simulating random images
 text_length = 100  # Arbitrary length for text tokens
-dummy_text_tokens = torch.randint(0, 32002, (batch_size, text_length))  # Simulating random text tokens
+dummy_text_tokens = torch.randint(0, 32002, (batch_size, text_length)).long()  # Simulating random text tokens
 
 # Pass the dummy data to the model
 outputs = model(text_tokens=dummy_text_tokens, images=dummy_images)
